OSIS-2800 Refactor delete
