+++ base/views/education_groups/delete.py
@@ -25,6 +25,7 @@
 ##############################################################################
 from django.urls import reverse_lazy
 
+import base.business.education_groups.delete
 from base.business.education_groups import shorten
 from base.models.education_group import EducationGroup
 from base.views.education_groups.perms import can_delete_all_education_group
@@ -56,7 +57,7 @@ class DeleteGroupEducationView(DeleteViewWithDependencies):
         self.education_group_years = self.get_object().educationgroupyear_set.all().order_by('academic_year__year')
         protected_messages = []
         for education_group_year in self.education_group_years:
-            protected_message = shorten.get_protected_messages_by_education_group_year(education_group_year)
+            protected_message = base.business.education_groups.delete.get_protected_messages_by_education_group_year(education_group_year)
             if protected_message:
                 protected_messages.append({
                     'education_group_year': education_group_year,
